+++ src/renderer/src/components/message/index.ts
@@ -3,4 +3,3 @@ export * from './ReflectedMessageBlock';
 export * from './CommandOutputMessageBlock';
 export * from './Messages';
 export * from './UserMessageBlock';
-export * from './InterruptedMessageBlock';