+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - added worktree directory important reminder when working in the worktree
 - added shift-click support for selecting task ranges in multiselect mode
 - correctly showing uncut list of uncommited files in the worktree status
+- added task actions based on the task state for easier workflow
 
 ## [0.46.0]